actionBar started
