--- src/assets/add.png:Zone.Identifier
@@ -1,4 +1,0 @@
-[ZoneTransfer]
-ZoneId=3
-ReferrerUrl=https://www.flaticon.com/search?word=plus%20icon
-HostUrl=https://www.flaticon.com/download/icon/3524384?icon_id=3524384&author=417&team=417&keyword=Plus&pack=packs%2Falphabet-and-numbers-9&style=4&format=png&color=%23000000&colored=1&size=512&selection=1&premium=0&type=standard&token=03AD1IbLBnTnibpIKaPXIVbP7SanZuQrW2yehrpo5zNSUWo_MrFftME9nq0O-5b7AWgeDwRkvFFO-ZVKCtvWKHl2PIew2RenOsZMVNYRpyHtYHj84g9qUArebPJITEu9o8wxKlmEhIOSvDdwtW3c1dBtilUDRi-QdlHHscOYN2d6IV-5FMAmoH5WRnDk6qz24hGIqWg8qlSjxydirQhAlVUyn1lAn5ICAkfp0v0jjCbi-oIAEDtEB6JGfsAJWFPM-hITahxwYxUj7kAIScVUv8HgAaaAQEDbW3xfU6EMknLbNfxIX977cTCQQoKjyOaQarjQJ5SSoGdlif9su0phJl2zu2AJC1C-niuXUJaVtup7-o9tGpOejQGSQcCpN8Xw2C8uV6Jvi5CyXJ6y1-d1OSSg4yjVNa6HoidObHk7OhzGI1gMA95OdopX6tBgXKWlE7HuFLtN1K7sVgzTGJxExL22YKWm8zj72pKsmH45zTE5GgXkt3-i-m63tihH_6YR4KS42uV1mw-jXqBdrqCMxR4cwC6FzvoMSFn5Uj1M_IToUA-uMOhYXDMziOo20_fIRj1WZZql4a9xll_bMM6XJDw48NgCW2PJ6Kzlj61l2tQrJjFv2YV3q5hiu2V4fH8w1U3k3Ic5cPsv9tm8vbpD8qctfRjfsvxwWeGB5LIqoZJ8HN0U3TZXIpfzfNVrrosh-d2k0jdY43bNb_8jwvm9HX5I1aLr8oAhsch3v8h-F1iZIragJPLJTO64Tn6zuMREzjv9v3YtnbrkBnDL6frJfnQd2hwSPzgibUvrnlEX6nvb2_n1-YoK2XqX_SLvPFWWswGcDrLNA2W_2PX-44iI5LMAp9dVOtgcgVao9gw2juMSMOfbJQn8kfPXQK7IB8DsqrHJXVyMNDl3laxfnEqAbzkfBnNCf0DQ_M_tGVl8VjZMa6ZkD_ZAQ3kssEjIhRdxxdGjvA78DAGpqYLyscu33VVI2Vk9qm4ry9IJhcUSSfAiHM_t9HU_i1bg_KgugfKf4JUIZhzyIB4PYCclDv8S1SZeTEVigTu2uRcLTr9ANv1mDw0Z-x0NOgqsxZNoZ1b_vunEdmSu-iDor5iRtjF-NPKhtI9J39Dc4Hm4Z_4Dx1oXIgzMakpTTfGoQaqh-BwQLWSLDEK-bFRk2Nwmaes8NkVfV37B5mhzeZAh8A8NLd6MjLYWtRFcEtROWv54RnCwzzxYBUuPO3iDyoohriwRWsaQtNyeDVayAMcoCoICEtqpvyHNIDdQJuDfzvnqNXzDWhshw5bvLHb5bv&search=plus+icon
