--- src/assets/addBlack.png:Zone.Identifier
@@ -1,4 +1,0 @@
-[ZoneTransfer]
-ZoneId=3
-ReferrerUrl=https://www.flaticon.com/free-icon/add_149705?term=plus&page=1&position=21&origin=search&related_id=149705
-HostUrl=https://www.flaticon.com/download/icon/149705?icon_id=149705&author=159&team=159&keyword=Add&pack=149629&style=Fill&style_id=27&format=png&color=&colored=1&size=512%2C256%2C128%2C64%2C32%2C24%2C16&selection=1&premium=&type=standard&search=plus
