--- src/assets/bubbleChat.png:Zone.Identifier
@@ -1,4 +1,0 @@
-[ZoneTransfer]
-ZoneId=3
-ReferrerUrl=https://www.flaticon.com/search?word=comments
-HostUrl=https://www.flaticon.com/download/icon/3031126?icon_id=3031126&author=741&team=741&keyword=Bubble+chat&pack=packs%2Finterface-180&style=1215&format=png&color=%23000000&colored=1&size=512&selection=1&premium=0&type=standard&search=comments
