--- src/assets/heart.png:Zone.Identifier
@@ -1,4 +1,0 @@
-[ZoneTransfer]
-ZoneId=3
-ReferrerUrl=https://www.flaticon.com/search?word=heart
-HostUrl=https://www.flaticon.com/download/icon/1077035?icon_id=1077035&author=256&team=256&keyword=Heart&pack=packs%2Fsocial-network-12&style=4&format=png&color=%23000000&colored=1&size=512&selection=1&premium=0&type=standard&token=03AD1IbLAkPptewsjWdINeJwQb7hgQUOG7_toOduvdUf6EIerDQl1pg7_jFg6zm4LPloa5k8gacUnjPA5vZGDjGZQ2oSUdoM7WLlD0dFd2D2_GB8aAenJyhyMXGcWmOwNu53BQJeevOmbSNr0WP_RGKV78vnefn8FUGJom8c6Hqe67n8UhYuRFVdV1JJSqMaC7p2JwBHJgJ0ruchGOtAkS-CPhRGFKwk4TIGja395KZWy2iL-1lFny1XySBZUSue8gG78EDnrTV8As1ulvmS3GvvZYcVUVNGbv5zCidnQxcCILZ2JjvbVrKTzEXOhoBwsNWDZodiACGrQGRUVvsYRxZmV4UQlVA5Fp-OyBbLxpHJzca1YdBImUhlmybWoTd4jq-BCXFK02BmhMZ5jTMv5FUwRhS-Mq_wlbkxYILJSxvbfDEk1Na2zxTnn0N7-YZu9Q3u2UoIk4_w4j5Z1IRvFHxh5K65Jv-Uff8Ue-MrfroW8vC0_bom-H8MwdbgDzSVv8s1pNWwcXXtzLP6hm3ArOswYVulxMIc3rShwqb_WuMFPEzwYHhZLFc77i1-VTHSPT9ycZMzLMZsbPtc_7R7ji-M2q1FWBKFDhMoxJVyMDUXsH9iApE03w3sL6XOvCyzsxGpOWkgDKh2cVzIM5TkjuW6wt44XMplau-n-Zpibknek-6WVS3Q6rYOtXl_-qttyi4KQLG43lMrBEAi5QIDudlFvRL_8tcUqWCQ7ro5fnBFhkK3T1YsdT6FFJu3hlcejuctZ9CVsLxJhvis6Yhip1Wlt5sTY9hoRKs6eP99GK_i5Ea32Gbvy1v394A3vaiChbof0FoUdyAQZMTLWjuryXlYojePVTbjSy-wO-BO13WkHUXSqYD191e3NnTavePoDUmNnbfRvgPIgA-2GsSLYkB4uLufDoi5zwqGCqAVMJsDTE9-a2TZd_Wh7BPE9_UlAkpYdKxj_PpYBZz1m9eMOsmKizRTTwm0m4qyP40n3SK-SEsMrlEIDH0wxGTWO249LZDwxQgxo2pNYUNXG2mGWdDficeDcpagdcv4DwhGKoRL3hUcxC0ftJXqUaPkVCwDrTmJ_KWvzuCWM4BGIP6NciPdsKV8cIr10Uf553QDnDzqumYS8t-gcX8mVcozSqR2Bdd6781eyXvR1ZcewkZJQ2fuNPxlfVwLQl3hjjCvNzBKvIcNEHjXJtFNVOYo0gCgeONLXFtr-f0lUTLK2AHc6BzIM6chJDI556M9mIvmhaineeB0IE8vdWV2PIjQcglR5pgKfxRXFOtBgD&search=heart
