--- src/assets/heartRed.png:Zone.Identifier
@@ -1,4 +1,0 @@
-[ZoneTransfer]
-ZoneId=3
-ReferrerUrl=https://www.flaticon.com/search?word=heart
-HostUrl=https://www.flaticon.com/download/icon/833472?icon_id=833472&author=321&team=321&keyword=Heart&pack=packs%2Fmiscellaneous-31&style=167&format=png&color=&colored=2&size=512&selection=1&premium=0&type=standard&search=heart
