--- src/assets/tickets.png:Zone.Identifier
@@ -1,4 +1,0 @@
-[ZoneTransfer]
-ZoneId=3
-ReferrerUrl=https://www.flaticon.com/search?word=tickets
-HostUrl=https://www.flaticon.com/download/icon/785581?icon_id=785581&author=174&team=174&keyword=Tickets&pack=packs%2Fevents-2&style=1&format=png&color=%23000000&colored=1&size=512&selection=1&premium=0&type=standard&search=tickets
