+++ src/hooks/useCounter.js
@@ -1,6 +1,6 @@
 import { useState } from "react"
 
-export const useCounter = (initialState = 10) => {
+export const useCounter = (initialState = 0) => {
  
     const [counter, setCounter] = useState(initialState)
