+++ src/styles/styles.scss
@@ -3,3 +3,4 @@
 
 @import './components/header';
 @import './components/home_card';
+@import './components/button'
